+++ apps/portal-web/src/pages/files/currentTransferInfo.tsx
@@ -14,17 +14,23 @@ import { NextPage } from "next";
 import { requireAuth } from "src/auth/requireAuth";
 import { PageTitle } from "src/components/PageTitle";
 import { Redirect } from "src/components/Redirect";
+import { prefix, useI18nTranslateToString } from "src/i18n";
 import { TransferInfoTable } from "src/pageComponents/filemanager/TransferInfoTable";
 import { publicConfig } from "src/utils/config";
 
+const p = prefix("pages.files.currentTransferInfo.");
+
 export const FileTransferPage: NextPage = requireAuth(() => true)(() => {
+
+  const t = useI18nTranslateToString();
+
   if (!publicConfig.CROSS_CLUSTER_FILE_TRANSFER_ENABLED) {
     return <Redirect url={"/dashboard"} />;
   }
 
   return (
     <div>
-      <PageTitle titleText={"文件传输进度查看"} />
+      <PageTitle titleText={t(p("checkTransfer"))} />
       <TransferInfoTable />
     </div>
   );