+++ apps/portal-server/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @scow/portal-server
 
+## 1.2.1
+
+## 1.2.0
+
+### Minor Changes
+
+- 135f2b1be3: 在门户系统的文件管理下，新增将文件直接作为作业文本提交调度器执行的功能，如果调度器 API 版本低于此接口版本报错
+
+### Patch Changes
+
+- af6a53dfcf: portal-server,auth,mis-server,audit-server 下 pino 日志的时间格式修改为八时区下的 YYYY-MM-DD HH:mm:ss
+- Updated dependencies [a3d2f44af6]
+- Updated dependencies [f42488eb9e]
+- Updated dependencies [5d2b75ccec]
+- Updated dependencies [a79aa109bb]
+- Updated dependencies [135f2b1be3]
+- Updated dependencies [5d2b75ccec]
+- Updated dependencies [f577d9d1e4]
+  - @scow/config@1.2.0
+  - @scow/lib-ssh@1.0.1
+  - @scow/scheduler-adapter-protos@1.1.0
+  - @scow/lib-scheduler-adapter@1.1.0
+  - @scow/utils@1.1.0
+  - @scow/lib-server@1.1.0
+  - @scow/protos@1.0.2
+  - @scow/rich-error-model@2.0.0
+
 ## 1.1.0
 
 ### Minor Changes