+++ libs/web/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @scow/lib-web
 
+## 1.2.1
+
+### Patch Changes
+
+- b84e4f9cc4: 暂时解决 tab 标签在最右侧时的抖动问题
+
+## 1.2.0
+
+### Minor Changes
+
+- 5d2b75ccec: 增加用户指定系统语言功能，可以指定系统唯一语言不再进行语言切换，也可以指定进入 SCOW 时的默认初始语言
+
+### Patch Changes
+
+- Updated dependencies [135f2b1be3]
+  - @scow/utils@1.1.0
+
 ## 1.1.0
 
 ### Minor Changes